Network.AWS

Retrieving Authorisation

data Credentials

Determines how authentication information is specified or retrieved.

Constructors

FromKeys AccessKey SecretKey
Explicit access and secret keys.

Note: you can achieve the same result purely by using fromKeys.

FromSession AccessKey SecretKey SecurityToken
A session containing the access key, secret key, and a security token.

Note: you can achieve the same result purely by using fromSession.

FromProfile ByteString
An IAM Profile name to lookup from the local EC2 instance-data.

FromEnv ByteString ByteString
Environment variables to lookup for the access and secret keys.

Discover
Attempt to read the default access and secret keys from the environment, falling back to the first available IAM profile if they are not set.

Note: This attempts to resolve http://instance-data rather than directly retrieving http://169.254.169.254 for IAM profile information to ensure the dns lookup terminates promptly if not running on EC2.

getAuth :: MonadIO m => Credentials -> EitherT Error m Auth

Environment Variables

accessKey

Arguments

:: ByteString
AWS_ACCESS_KEY

Default access key environment variable.

secretKey

Arguments

:: ByteString
AWS_SECRET_KEY

Default secret key environment variable.

Sending Requests

Synchronous

send

Arguments

:: (MonadResource m, AWSRequest a, AWSSigner (Signer' (Service' a)))

=> Auth
AWS authentication credentials.

-> Region
AWS Region.

-> a
Request to send.

-> Manager
HTTP Manager.

-> m (Either (Error' (Service' a)) (Response' a))

Pagination

paginate

Arguments

:: (MonadResource m, AWSPager a, AWSSigner (Signer' (Service' a)))

=> Auth
AWS authentication credentials.

-> Region
AWS Region.

-> a
Seed request to send.

-> Manager
HTTP Manager.

-> Source m (Either (Error' (Service' a)) (Response' a))

Signed URLs

presign

Arguments

:: (AWSRequest a, AWSPresigner (Signer' (Service' a)))

=> Auth
AWS authentication credentials.

-> Region
AWS Region.

-> a
Request to presign.

-> Int
Expiry time in seconds.

-> UTCTime
Signing time.

-> Signed a (Signer' (Service' a))