Network.AWS.Auth

Description

Loading your Amazon AWS credentials.

Defaults

accessKey

Arguments

:: ByteString
AWS_ACCESS_KEY

Default access key environment variable. See: fromEnv

secretKey

Arguments

:: ByteString
AWS_SECRET_KEY

Default secret key environment variable. See: fromEnv

Specifying credentials

fromKeys :: AccessKey -> SecretKey -> Auth

Explicit access and secret keys.

fromSession :: AccessKey -> SecretKey -> SecurityToken -> Auth

A session containing the access key, secret key, and a security token.

discover :: MonadIO m => EitherT Error m Auth

Attempt to read the default access and secret keys from the environment, falling back to the first available IAM profile if they are not set.

Note: This attempts to resolve http://instance-data rather than directly retrieving http://169.254.169.254 for IAM profile information to ensure the dns lookup terminates promptly if not running on EC2.

fromEnv :: MonadIO m => EitherT Error m Auth

Retrieve access and secret keys from the default environment variables.

See: accessKey and secretKey

fromEnv' :: MonadIO m => ByteString -> ByteString -> EitherT Error m Auth

Retrieve access and secret keys from specific environment variables.

fromProfile :: MonadIO m => EitherT Error m Auth

Retrieve the default IAM Profile from the local EC2 instance-data.

This determined by Amazon as the first IAM profile found in the response from: http://169.254.169.254/latest/meta-data/iam/security-credentials/

fromProfile' :: MonadIO m => ByteString -> EitherT Error m Auth

Lookup a specific IAM Profile by name from the local EC2 instance-data.